refactoring #1: got rid of config field in Store
- moved store code to sqlstore

closes #1
