+++ internal/app/store/sqlstore/store_test.go
@@ -1,4 +1,4 @@
-package store_test
+package sqlstore_test
 
 import (
 	"os"
