+++ internal/app/store/sqlstore/userrepository_test.go
@@ -1,16 +1,16 @@
-package store_test
+package sqlstore_test
 
 import (
 	"github.com/SharpDevOps10/GoPractice/internal/app/model"
-	"github.com/SharpDevOps10/GoPractice/internal/app/store"
+	"github.com/SharpDevOps10/GoPractice/internal/app/store/sqlstore"
 	"github.com/stretchr/testify/assert"
 	"testing"
 )
 
 func TestUserRepository_Create(t *testing.T) {
-	db, teardown := store.TestDB(t, databaseURL)
+	db, teardown := sqlstore.TestDB(t, databaseURL)
 	defer teardown("userss")
-	s := store.New(db)
+	s := sqlstore.New(db)
 
 	u, err := s.User().Create(model.TestUser(t))
 	assert.NoError(t, err)
@@ -18,10 +18,10 @@ func TestUserRepository_Create(t *testing.T) {
 }
 
 func TestUserRepository_FindByEmail(t *testing.T) {
-	db, teardown := store.TestDB(t, databaseURL)
+	db, teardown := sqlstore.TestDB(t, databaseURL)
 	defer teardown("userss")
 
-	s := store.New(db)
+	s := sqlstore.New(db)
 	email := "bulldog@gmail.com"
 	_, err := s.User().FindByEmail(email)
 	assert.Error(t, err)
